Move `OntologyTypeVertxId` to `utoipa_typedef`
