+++ apps/hash-graph/lib/graph/src/api/rest.rs
@@ -37,8 +37,8 @@ use crate::{
         middleware::log_request_and_response,
         utoipa_typedef::subgraph::{
             Edges, KnowledgeGraphOutwardEdge, KnowledgeGraphRootedEdges, KnowledgeGraphVertex,
-            KnowledgeGraphVertices, OntologyOutwardEdge, OntologyRootedEdges, OntologyVertex,
-            OntologyVertices, Subgraph, Vertex, Vertices,
+            KnowledgeGraphVertices, OntologyOutwardEdge, OntologyRootedEdges, OntologyTypeVertexId,
+            OntologyVertex, OntologyVertices, Subgraph, Vertex, Vertices,
         },
     },
     identifier::{
@@ -62,7 +62,7 @@ use crate::{
         },
         identifier::{
             DataTypeVertexId, EntityIdWithInterval, EntityTypeVertexId, EntityVertexId,
-            GraphElementVertexId, OntologyTypeVertexId, PropertyTypeVertexId,
+            GraphElementVertexId, PropertyTypeVertexId,
         },
         temporal_axes::{QueryTemporalAxes, QueryTemporalAxesUnresolved, SubgraphTemporalAxes},
     },
