+++ apps/hash-graph/lib/graph/src/subgraph/identifier.rs
@@ -5,6 +5,6 @@ pub use self::{
     edge::{EdgeEndpoint, EntityIdWithInterval},
     vertex::{
         DataTypeVertexId, EntityTypeVertexId, EntityVertexId, GraphElementVertexId,
-        OntologyTypeVertexId, PropertyTypeVertexId, VertexId,
+        PropertyTypeVertexId, VertexId,
     },
 };
